DROID-2378 move membership model
