+++ payments/src/main/java/com/anytypeio/anytype/payments/screens/TierPreviewView.kt
@@ -46,7 +46,7 @@ import com.anytypeio.anytype.payments.models.Tier
 import com.anytypeio.anytype.payments.models.TierConditionInfo
 import com.anytypeio.anytype.payments.models.TierPeriod
 import com.anytypeio.anytype.presentation.editor.cover.CoverColor
-import com.anytypeio.anytype.presentation.membership.models.TierId
+import com.anytypeio.anytype.core_models.membership.TierId
 
 @Composable
 fun TierPreviewView(
