+++ payments/src/main/java/com/anytypeio/anytype/payments/viewmodel/MembershipViewModel.kt
@@ -34,8 +34,8 @@ import com.anytypeio.anytype.payments.playbilling.BillingPurchaseState
 import com.anytypeio.anytype.presentation.extension.sendAnalyticsMembershipClickEvent
 import com.anytypeio.anytype.presentation.extension.sendAnalyticsMembershipPurchaseEvent
 import com.anytypeio.anytype.presentation.extension.sendAnalyticsMembershipScreenEvent
-import com.anytypeio.anytype.presentation.membership.models.MembershipStatus
-import com.anytypeio.anytype.presentation.membership.models.TierId
+import com.anytypeio.anytype.core_models.membership.MembershipStatus
+import com.anytypeio.anytype.core_models.membership.TierId
 import com.anytypeio.anytype.presentation.membership.provider.MembershipProvider
 import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.Job
