+++ payments/src/test/java/com/anytypeio/anytype/payments/TierActiveAndFreeTests.kt
@@ -15,8 +15,8 @@ import com.anytypeio.anytype.payments.models.TierPreview
 import com.anytypeio.anytype.payments.playbilling.BillingClientState
 import com.anytypeio.anytype.payments.viewmodel.MembershipMainState
 import com.anytypeio.anytype.payments.viewmodel.MembershipTierState
-import com.anytypeio.anytype.presentation.membership.models.MembershipStatus
-import com.anytypeio.anytype.presentation.membership.models.TierId
+import com.anytypeio.anytype.core_models.membership.MembershipStatus
+import com.anytypeio.anytype.core_models.membership.TierId
 import junit.framework.TestCase
 import kotlin.test.assertIs
 import kotlinx.coroutines.test.runTest
